AEventManager - Store total momentum/energy of a combination as a double instead of an integer. This fixes the photon combination problem.

+++ trunk/src/aeventmanager/ATrackCombination.h
@@ -39,10 +39,11 @@ class AEVENTMANAGER_EXPORT ATrackCombination : public ATrack
  private:
   QList<ATrack*> tracks;
 
-  int Px;
-  int Py;
-  int Pz;
-  int E;
+  double Px;
+  double Py;
+  double Pz;
+  double E;
+  
   float InvMass;
 
   void recalculate();
